refactor: Move utils.typing into utils.common.typing
