+++ utils/common/typing.py
@@ -1,3 +1,4 @@
+"""Utility functions for dealing with Python generic types."""
 from typing import get_args, get_origin
 import types
